It is considered the case of using int32 for namespace and int32 for command. When a message is constructed, it will use int64 for the full code if BOTH namespace and command are integer arguments. Otherwise it will use strings as before.
When serializing a Message, it will not include, anymore, the args and kwargs elements if they are empty. This is to optimize the protocols bandwidth.

When unserializing, the code will be allowed to be an int/long value, and will be split in a different way. Args and kwargs are contemplated to be not present, as well.

When sending a message, an additional parameter is specified (a named parameter `use_msgpack`) to allow the user to pass False to it, so they can fall back to a JSON protocol. If such value is not present, the value in `use_msgpack` property will be considered.

When receiving a message, it will be checked if it starts with '{'. In such way, the message will be treated (and tried) as JSON. Otherwise if will be treated and tried as a msgpack map.

The package will be edited to support these changes and the dependency to msgpack-python>=0.4.6